requirements.txt
# DART API Access
dart-fss>=0.3.10

# Azure SDKs
azure-ai-documentintelligence>=1.0.0b3
azure-search-documents>=11.4.0
azure-identity>=1.15.0

# OpenAI API
openai>=1.3.0

# Data Handling & Numerics
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# HTML/XML Parsing
beautifulsoup4>=4.9.3
lxml>=4.6.3

# Configuration & Utilities
python-dotenv>=0.19.0
tenacity>=8.2.0

# Text Splitter (Optional, if using LangChain)
langchain-text-splitters>=0.0.1

# Development Tools
ipython
jupyter
